+++ Projects/Server/Json/JsonConfig.cs
@@ -49,6 +49,7 @@ public static class JsonConfig
         options.Converters.Add(new IPEndPointConverterFactory());
         options.Converters.Add(new TypeConverterFactory());
         options.Converters.Add(new WorldLocationConverterFactory());
+        options.Converters.Add(new TextDefinitionConverterFactory());
 
         for (var i = 0; i < converters.Length; i++)
         {
